🚀  Refactoring common requests that return lists
