--- backend/abba/src/main/java/com/zon/abba/member/request/member/MemberListRequest.java
@@ -1,19 +1,0 @@
-package com.zon.abba.member.request.member;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class MemberListRequest {
-    private Integer pageNo;
-    private Integer pageSize;
-    private String sort;
-    private String sortValue;
-    private String filter;
-    private String filterValue;
-}
