--- backend/abba/src/main/java/com/zon/abba/member/request/recommend/ListRecommendRequest.java
@@ -1,15 +1,0 @@
-package com.zon.abba.member.request.recommend;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ListRecommendRequest {
-    private Integer pageNo;
-    private Integer pageSize;
-}
